Rename path case/[id] to incidents/[id]
