+++ __e2e__/cases.spec.ts
@@ -17,7 +17,7 @@ test("Test the Cases Page", async ({ page }) => {
   await expect(
     casesPage.tableHeaders.locator("text=date received")
   ).toHaveCount(1)
-  await expect(casesPage.tableHeaders.locator("text=case status")).toHaveCount(
+  await expect(casesPage.tableHeaders.locator("text=status")).toHaveCount(
     1
   )
   await expect(
@@ -33,6 +33,6 @@ test("Test the Cases Page", async ({ page }) => {
     page.waitForNavigation(),
   ])
 
-  // Should be on '/case/<caseID>' page after clicking a row
-  await expect(page.url()).toContain("/case/1")
+  // Should be on '/incidents/<incidentID>' page after clicking a row
+  await expect(page.url()).toContain("/incidents/1")
 })
